buttons changed to image and problems discovered
ok so the problem with the add button is the fact that whatever way the
id="create-customer" is configured it will only work for the first
button, i cant work out why. also i noticed that when i reload the page
it sometimes adds more customer segments, like it adds the last one i
added again.